+++ tests/testthat/test-cache.R
@@ -387,7 +387,7 @@ test_that("test objSize", {
 
   a <- simInit(objects = list(d = 1:10, b = 2:20))
   os <- objSize(a)
-  expect_true(length(os) == 6) # 4 objects, the environment, the rest
+  expect_true(length(os) == 2)
 })
 
 test_that("Cache sim objs via .Cache attr", {