+++ tests/testthat/test-paths.R
@@ -19,7 +19,8 @@ test_that("paths file does not work correctly", {
                  modulePath = paths$modulePath,
                  outputPath = getPaths()[["outputPath"]],
                  rasterPath = getPaths()[["rasterPath"]],
-                 scratchPath = getPaths()[["scratchPath"]]
+                 scratchPath = getPaths()[["scratchPath"]],
+                 terraPath = getPaths()[["terraPath"]]
                 ), normPath)
               )
 
@@ -33,13 +34,15 @@ test_that("paths file does not work correctly", {
                            modulePath = paths$modulePath,
                            outputPath = path.expand(paths$outputPath),
                            rasterPath = getPaths()[["rasterPath"]],
-                           scratchPath = getPaths()[["scratchPath"]]
+                           scratchPath = getPaths()[["scratchPath"]],
+                           terraPath = getPaths()[["terraPath"]]
                ), normPath))
 
   # test for all unnamed
   paths <- list(tmpdir,
                 tmpdir,
                 system.file("sampleModules", package = "SpaDES.core"),
+                tmpdir,
                 tmpdir,
                 tmpdir,
                 tmpdir)
@@ -50,16 +53,20 @@ test_that("paths file does not work correctly", {
                            modulePath = paths[[3]],
                            outputPath = paths[[4]],
                            rasterPath = paths[[5]],
-                           scratchPath = paths[[6]]
+                           scratchPath = paths[[6]],
+                           terraPath = paths[[7]]
                ), normPath))
 
   # test for all named, non consecutive, using accessors
-  paths <- list(cachePath = tmpdir,
-                modulePath = system.file("sampleModules", package = "SpaDES.core"),
-                outputPath = tmpdir,
-                inputPath = tmpdir,
-                rasterPath = file.path(tmpdir, "raster"), ## subdir of scratchPath
-                scratchPath = tmpdir)
+  paths <- list(
+    cachePath = tmpdir,
+    modulePath = system.file("sampleModules", package = "SpaDES.core"),
+    outputPath = tmpdir,
+    inputPath = tmpdir,
+    rasterPath = file.path(tmpdir, "raster"), ## subdir of scratchPath
+    scratchPath = tmpdir,
+    terraPath = file.path(tmpdir, "terra") ## subdir of scratchPath
+  )
   mySim <- simInit(times, params, modules, objects = list(), paths)
   expect_equal(lapply(paths(mySim), normPath),
                lapply(list(cachePath = cachePath(mySim),
@@ -67,7 +74,8 @@ test_that("paths file does not work correctly", {
                            modulePath = modulePath(mySim),
                            outputPath = outputPath(mySim),
                            rasterPath = rasterPath(mySim),
-                           scratchPath = scratchPath(mySim)
+                           scratchPath = scratchPath(mySim),
+                           terraPath = terraPath(mySim)
                ), normPath))
 
   # missing paths
@@ -99,4 +107,7 @@ test_that("paths file does not work correctly", {
 
   scratchPath(mySim) <- tmpdir
   expect_equal(scratchPath(mySim), tmpdir)
+
+  terraPath(mySim) <- tmpdir
+  expect_equal(terraPath(mySim), tmpdir)
 })